refactor: set up ts with tsconfig.ts refactor: use image source map refactor: use sass refactor: move components into own folder
