+++ .gitignore
@@ -5,6 +5,7 @@
 /ressources
 /src/ressources
 /public/res
+/public/images
 /.pnp
 .pnp.js
