+++ package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "fs": "^0.0.1-security",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -13,6 +15,7 @@
     "react-scripts": "^5.0.1",
     "react-timeline-animation": "^1.2.3",
     "sass": "^1.69.5",
+    "typescript": "^4.5.4",
     "web-vitals": "^3.5.0"
   },
   "scripts": {
