+++ src/components/BottomBar/BottomBar.scss
