+++ src/components/CopyRight/CopyRight.scss
