+++ src/components/Texts/Texts.scss
